TAP5-1881: TypeCoercion from Number to Boolean returns false for any number that is an even multiple of 256
git-svn-id: https://svn.apache.org/repos/asf/tapestry/tapestry5/trunk@1303549 13f79535-47bb-0310-9956-ffa450edef68

+++ tapestry-ioc/src/main/java/org/apache/tapestry5/ioc/services/TapestryIOCModule.java
@@ -310,7 +310,7 @@ public final class TapestryIOCModule
         {
             public Boolean coerce(Number input)
             {
-                return input.byteValue() != 0;
+                return input.longValue() != 0;
             }
         });
